Move IconButton to ProfilePage root, refactored UserPhoto
