+++ web-client/src/components/ProfilePage/UserPhoto/index.module.css
